Cleanup
Fixed landscape keypad taps causing the lockscreen to jump up and down.
Fixed the rounding issue introduced in the last commit.
Minor touchups here and there.

+++ src/window.hpp
@@ -27,7 +27,7 @@ class syslock : public Gtk::Window {
 
 	private:
 		bool locked;
-		int window_height;
+		double window_height;
 		int start_height;
 		bool auth;
 		sigc::connection connection;
